moved generated API clients (except cBioPortal API client) under public lib
