+++ commons.webpack.config.js
@@ -7,7 +7,7 @@ var GeneratePackageJsonPlugin =require('generate-package-json-webpack-plugin');
 var TerserPlugin = require('terser-webpack-plugin');
 
 var commit = 'unknown';
-var version = '0.0.13';
+var version = '0.0.14';
 
 const NODE_ENV = process.env.NODE_ENV || 'development';
