+++ package.json
@@ -16,14 +16,14 @@
     "fetchAPI": "./scripts/env_vars.sh && eval \"$(./scripts/env_vars.sh)\" && curl ${CBIOPORTAL_URL}/api/api-docs | json | grep -v basePath | grep -v termsOfService | grep -v host > src/shared/api/generated/CBioPortalAPI-docs.json && curl ${CBIOPORTAL_URL}/api/api-docs?group=internal | json | grep -v host | grep -v basePath | grep -v termsOfService > src/shared/api/generated/CBioPortalAPIInternal-docs.json",
     "buildAPI": "node scripts/generate-api.js src/shared/api/generated CBioPortalAPI CBioPortalAPIInternal",
     "updateOncoKbAPI": "yarn run fetchOncoKbAPI && yarn run buildOncoKbAPI",
-    "fetchOncoKbAPI": "curl https://www.oncokb.org/api/v1/v2/api-docs | json | grep -v basePath | grep -v termsOfService | grep -v host > src/shared/api/generated/OncoKbAPI-docs.json",
-    "buildOncoKbAPI": "node scripts/generate-api.js src/shared/api/generated OncoKbAPI",
+    "fetchOncoKbAPI": "curl https://www.oncokb.org/api/v1/v2/api-docs | json | grep -v basePath | grep -v termsOfService | grep -v host > src/public-lib/api/generated/OncoKbAPI-docs.json",
+    "buildOncoKbAPI": "node scripts/generate-api.js src/public-lib/api/generated OncoKbAPI",
     "updateG2SAPI": "yarn run fetchG2SAPI && yarn run buildG2SAPI",
-    "fetchG2SAPI": "curl http://g2s.genomenexus.org/v2/api-docs?group=api > src/shared/api/generated/Genome2StructureAPI-docs.json",
-    "buildG2SAPI": "node scripts/generate-api.js src/shared/api/generated Genome2StructureAPI",
+    "fetchG2SAPI": "curl http://g2s.genomenexus.org/v2/api-docs?group=api > src/public-lib/api/generated/Genome2StructureAPI-docs.json",
+    "buildG2SAPI": "node scripts/generate-api.js src/public-lib/api/generated Genome2StructureAPI",
     "updateGenomeNexusAPI": "yarn run fetchGenomeNexusAPI && yarn run buildGenomeNexusAPI",
-    "fetchGenomeNexusAPI": "./scripts/env_vars.sh && eval \"$(./scripts/env_vars.sh)\" && curl ${GENOME_NEXUS_URL}/v2/api-docs | json | grep -v basePath | grep -v termsOfService | grep -v host > src/shared/api/generated/GenomeNexusAPI-docs.json && curl ${GENOME_NEXUS_URL}/v2/api-docs?group=internal | json | grep -v basePath | grep -v termsOfService | grep -v host > src/shared/api/generated/GenomeNexusAPIInternal-docs.json",
-    "buildGenomeNexusAPI": "node scripts/generate-api.js src/shared/api/generated GenomeNexusAPI GenomeNexusAPIInternal",
+    "fetchGenomeNexusAPI": "./scripts/env_vars.sh && eval \"$(./scripts/env_vars.sh)\" && curl ${GENOME_NEXUS_URL}/v2/api-docs | json | grep -v basePath | grep -v termsOfService | grep -v host > src/public-lib/api/generated/GenomeNexusAPI-docs.json && curl ${GENOME_NEXUS_URL}/v2/api-docs?group=internal | json | grep -v basePath | grep -v termsOfService | grep -v host > src/public-lib/api/generated/GenomeNexusAPIInternal-docs.json",
+    "buildGenomeNexusAPI": "node scripts/generate-api.js src/public-lib/api/generated GenomeNexusAPI GenomeNexusAPIInternal",
     "updateHotspotGenes": "./scripts/get_hotspot_genes.sh > src/shared/static-data/hotspotGenes.json",
     "compileOqlParser": "cd src/shared/lib/oql && pegjs oql-parser.pegjs",
     "serveDist": "./scripts/serve_dist.sh",
@@ -31,8 +31,8 @@
     "test": "export NODE_OPTIONS=--max-old-space-size=2048; karma start karma.conf.js",
     "test:watch": "yarn run test -- --watch",
     "test:debug": "yarn run test -- --debug --browsers=Chrome_with_debugging --single-run=false",
-    "lint": "tslint 'src/**/*.ts' 'src/**/*.tsx' --exclude 'src/shared/api/generated/**'",
-    "lint-fix": "tslint --fix 'src/**/*.ts' 'src/**/*.tsx' --exclude 'src/shared/api/generated/**'",
+    "lint": "tslint 'src/**/*.ts' 'src/**/*.tsx' --exclude 'src/shared/api/generated/**' --exclude 'src/public-lib/api/generated/**'",
+    "lint-fix": "tslint --fix 'src/**/*.ts' 'src/**/*.tsx' --exclude 'src/shared/api/generated/**' --exclude 'src/public-lib/api/generated/**'",
     "eslint": "eslint src/** --ext .js --ext .jsx --cache",
     "syncmock": "node src/test/fetchMockData.js --diff"
   },
