+++ src/pages/resultsView/ResultsViewPageStore.ts
@@ -93,7 +93,7 @@ import {
     GenesetMolecularData
 } from "../../shared/api/generated/CBioPortalAPIInternal";
 import internalClient from "../../shared/api/cbioportalInternalClientInstance";
-import {CancerGene, IndicatorQueryResp} from "../../shared/api/generated/OncoKbAPI";
+import {CancerGene, IndicatorQueryResp} from "../../public-lib/api/generated/OncoKbAPI";
 import {getAlterationString} from "../../shared/lib/CopyNumberUtils";
 import memoize from "memoize-weak-decorator";
 import request from "superagent";
@@ -128,7 +128,7 @@ import ClinicalDataCache, {
 import {getDefaultMolecularProfiles} from "../../shared/lib/getDefaultMolecularProfiles";
 import {getProteinPositionFromProteinChange} from "public-lib/lib/ProteinChangeUtils";
 import {isMutation} from "../../shared/lib/CBioPortalAPIUtils";
-import {VariantAnnotation} from "shared/api/generated/GenomeNexusAPI";
+import {VariantAnnotation} from "public-lib/api/generated/GenomeNexusAPI";
 import {ServerConfigHelpers} from "../../config/config";
 import {
     populateSampleSpecificationsFromVirtualStudies, ResultsViewTab,
