+++ src/pages/staticPages/tools/oncoprinter/OncoprinterUtils.ts
@@ -9,7 +9,7 @@ import {cna_profile_data_to_string} from "../../../../shared/lib/oql/AccessorsFo
 import {fillGeneticTrackDatum, OncoprintMutationType} from "../../../../shared/components/oncoprint/DataUtils";
 import {Gene, Mutation, NumericGeneMolecularData} from "../../../../shared/api/generated/CBioPortalAPI";
 import {getProteinPositionFromProteinChange} from "public-lib/lib/ProteinChangeUtils";
-import OncoKbAPI, {Query} from "../../../../shared/api/generated/OncoKbAPI";
+import OncoKbAPI, {Query} from "../../../../public-lib/api/generated/OncoKbAPI";
 import {ONCOKB_DEFAULT, queryOncoKbData} from "../../../../shared/lib/StoreUtils";
 import {generateQueryVariant} from "../../../../shared/lib/OncoKbUtils";
 import {generateQueryVariantId} from "public-lib/lib/OncoKbUtils";
