+++ src/shared/api/genomeNexusInternalClientInstance.ts
@@ -1,4 +1,4 @@
-import GenomeNexusAPIInternal from "./generated/GenomeNexusAPIInternal";
+import GenomeNexusAPIInternal from "../../public-lib/api/generated/GenomeNexusAPIInternal";
 
 const client = new GenomeNexusAPIInternal();
