+++ src/shared/cache/GenomeNexusCache.ts
@@ -1,6 +1,6 @@
 import {fetchVariantAnnotationsByMutation} from "shared/lib/StoreUtils";
 import {extractGenomicLocation, genomicLocationString} from "shared/lib/MutationUtils";
-import {VariantAnnotation} from "shared/api/generated/GenomeNexusAPI";
+import {VariantAnnotation} from "public-lib/api/generated/GenomeNexusAPI";
 import {Mutation} from "shared/api/generated/CBioPortalAPI";
 import LazyMobXCache, {CacheData} from "shared/lib/LazyMobXCache";
 import AppConfig from "appConfig";
