+++ src/shared/cache/ResidueMappingCache.ts
@@ -1,4 +1,4 @@
-import {default as Genome2StructureAPI, Alignment, ResidueMapping} from "shared/api/generated/Genome2StructureAPI";
+import {default as Genome2StructureAPI, Alignment, ResidueMapping} from "public-lib/api/generated/Genome2StructureAPI";
 import g2sClient from "shared/api/g2sClientInstance";
 import {CacheData} from "shared/lib/LazyMobXCache";
 import {remoteData} from "public-lib/api/remoteData";
