+++ src/shared/components/annotation/genomeNexus/MutationAssessor.tsx
@@ -3,7 +3,7 @@ import DefaultTooltip from 'public-lib/components/defaultTooltip/DefaultTooltip'
 import annotationStyles from "./../styles/annotation.module.scss";
 import classNames from 'classnames';
 import tooltipStyles from "./styles/mutationAssessorTooltip.module.scss";
-import {MutationAssessor as MutationAssessorData} from 'shared/api/generated/GenomeNexusAPIInternal';
+import {MutationAssessor as MutationAssessorData} from 'public-lib/api/generated/GenomeNexusAPIInternal';
 import mutationAssessorColumn from "./styles/mutationAssessorColumn.module.scss";
 
 export interface IMutationAssessorProps {
