+++ src/shared/components/mutationTable/column/AnnotationColumnFormatter.tsx
@@ -16,7 +16,7 @@ import {IOncoKbCancerGenesWrapper, IOncoKbData, IOncoKbDataWrapper} from "shared
 import {IMyCancerGenomeData, IMyCancerGenome} from "shared/model/MyCancerGenome";
 import {IHotspotDataWrapper} from "shared/model/CancerHotspots";
 import {CancerStudy, Mutation} from "shared/api/generated/CBioPortalAPI";
-import {CancerGene, IndicatorQueryResp, Query} from "shared/api/generated/OncoKbAPI";
+import {CancerGene, IndicatorQueryResp, Query} from "public-lib/api/generated/OncoKbAPI";
 import {getEvidenceQuery} from "shared/lib/OncoKbUtils";
 import {generateQueryVariantId} from "public-lib/lib/OncoKbUtils";
 import {is3dHotspot, isRecurrentHotspot} from "shared/lib/AnnotationUtils";
