+++ src/shared/components/mutationTable/column/GnomadColumnFormatter.tsx
@@ -7,7 +7,7 @@ import generalStyles from "./styles.module.scss";
 import {Mutation} from "shared/api/generated/CBioPortalAPI";
 import DefaultTooltip from 'public-lib/components/defaultTooltip/DefaultTooltip';
 import {default as TableCellStatusIndicator, TableCellStatus} from "public-lib/components/TableCellStatus";
-import {MyVariantInfo, MyVariantInfoAnnotation, Gnomad, AlleleCount, AlleleNumber, Homozygotes, AlleleFrequency} from 'shared/api/generated/GenomeNexusAPI';
+import {MyVariantInfo, MyVariantInfoAnnotation, Gnomad, AlleleCount, AlleleNumber, Homozygotes, AlleleFrequency} from 'public-lib/api/generated/GenomeNexusAPI';
 import GenomeNexusMyVariantInfoCache, { GenomeNexusCacheDataType } from "shared/cache/GenomeNexusMyVariantInfoCache";
 import GnomadFrequencyTable from 'shared/components/gnomad/GnomadFrequencyTable';
