+++ src/shared/lib/AnnotationUtils.ts
@@ -4,7 +4,7 @@ import {CosmicMutation} from "shared/api/generated/CBioPortalAPIInternal";
 import {ICosmicData} from "shared/model/Cosmic";
 import {IMyCancerGenome, IMyCancerGenomeData} from "shared/model/MyCancerGenome";
 import {Mutation} from "shared/api/generated/CBioPortalAPI";
-import {Hotspot} from "../api/generated/GenomeNexusAPI";
+import {Hotspot} from "../../public-lib/api/generated/GenomeNexusAPI";
 
 /**
  * Utility functions related to annotation data.
