+++ src/shared/lib/PdbUtils.ts
@@ -1,6 +1,6 @@
 import * as _ from 'lodash';
-import {PdbHeader} from "shared/api/generated/GenomeNexusAPI";
-import {Alignment} from "shared/api/generated/Genome2StructureAPI";
+import {PdbHeader} from "public-lib/api/generated/GenomeNexusAPI";
+import {Alignment} from "public-lib/api/generated/Genome2StructureAPI";
 import {
     IPdbPositionRange, IPdbChain, PdbAlignmentIndex,
     ALIGNMENT_GAP, ALIGNMENT_MINUS, ALIGNMENT_PLUS, ALIGNMENT_SPACE,
