+++ src/test/check_api_sync.sh
@@ -31,11 +31,11 @@ run_and_check_diff 'yarn run fetchAPI' 'src/shared/api/generated/CBioPortalAPI-d
 if [[ $? -gt 0 ]]; then
     sync_error_count=$(($sync_error_count + 1))
 fi
-run_and_check_diff 'yarn run fetchOncoKbAPI' src/shared/api/generated/OncoKbAPI-docs.json "${OUT_OF_SYNC_MSG}"
+run_and_check_diff 'yarn run fetchOncoKbAPI' src/public-lib/api/generated/OncoKbAPI-docs.json "${OUT_OF_SYNC_MSG}"
 if [[ $? -gt 0 ]]; then
     sync_error_count=$(($sync_error_count + 1))
 fi
-run_and_check_diff 'yarn run fetchGenomeNexusAPI' src/shared/api/generated/GenomeNexusAPI-docs.json "${OUT_OF_SYNC_MSG}"
+run_and_check_diff 'yarn run fetchGenomeNexusAPI' src/public-lib/api/generated/GenomeNexusAPI-docs.json "${OUT_OF_SYNC_MSG}"
 if [[ $? -gt 0 ]]; then
     sync_error_count=$(($sync_error_count + 1))
 fi
@@ -46,12 +46,12 @@ run_and_check_diff 'yarn run buildAPI' 'src/shared/api/generated/CBioPortalAPI.t
 if [[ $? -gt 0 ]]; then
     generation_error_count=$(($generation_error_count + 1))
 fi
-run_and_check_diff 'yarn run buildOncoKbAPI' src/shared/api/generated/OncoKbAPI.ts "${TS_GEN_MSG}"
+run_and_check_diff 'yarn run buildOncoKbAPI' src/public-lib/api/generated/OncoKbAPI.ts "${TS_GEN_MSG}"
 if [[ $? -gt 0 ]]; then
     generation_error_count=$(($generation_error_count + 1))
 fi
 exit $(($generation_error_count + $sync_error_count))
-run_and_check_diff 'yarn run buildGenomeNexusAPI' src/shared/api/generated/GenomeNexusAPI.ts "${TS_GEN_MSG}"
+run_and_check_diff 'yarn run buildGenomeNexusAPI' src/public-lib/api/generated/GenomeNexusAPI.ts "${TS_GEN_MSG}"
 if [[ $? -gt 0 ]]; then
     generation_error_count=$(($generation_error_count + 1))
 fi
